Refactor exception handling
Use three board exception types:
ConfigError  - for exceptions raised while processing command line
               arguments or config files
FileError    - for exceptions raised due to missing or duplicated
               class and node files.
ProcessError - for exceptions raised during value interpolation

When exceptions are raised they are caught multiple times then reraised
with addition information added. For example when a NoSuchReference error
is raised the the path and file of the key handing the reference is known
as in the key name of the missing reference. The exception is then caught
by the ParametersResolver and the hierarchy_type is value is added to the
exception and reraised. This repeats to add the node name to the exception.