Raise error on a recursive class include
When loading a class the included classes of a class are loaded
first, before the full load processing of the parent class. This
allows parameters set in included classes to be overriden by the
same parameters set in the parent class.

However this means it is possible that during a chain of includes to
loop back over to a class already included, but not fully processed.
In the original reclass this would cause an infinite recursion of
includes until python raised a recursion error. In nodeclass due to
the different processing of class loads the recursively included
class would be ignored the second time it was incountered in the same
way a class already fully loaded would be ignored on it's subsequent
inclusions.

These recursive includes are now disallowed and if found an
exception is raised.