Fix handling of failed inventory queries causing exceptions
Previously if a node's parameter data was calculated as part of an
inventory query and the node had a bad reference or included a
nonexistant class then the inventory query would fail and raise an
exception, irrespective of if the IgnoreErrors flag was set of not.

Fix this behaviour and correctly ignore errors for broken nodes if
the ignore errors flag is set. Instead of raising an exception a
warning is printed to stderr.